pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitlab-users"
description = "Export GitLab users information and automate user accounts creation"
readme = "README.md"
authors = [
    { name = "Matthieu Boileau", email = "matthieu.boileau@math.unistra.fr" },
]
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = ["python-gitlab>=5.0.0"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "ruff",
    "black",
    "pytest",
    "pytest-cov",
    "build"
]

[project.urls]
Homepage = "https://github.com/boileaum/gitlab-users"

[tool.setuptools.dynamic]
version = { attr = "gitlab_users.__version__" }

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
gitlab-users = "gitlab_users.cli:main"

[tool.black]
line-length = 88
target-version = ['py37']

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src/gitlab_users --cov-report=term-missing --cov-report=xml"
